+++ src/forms/index.js
@@ -15,6 +15,7 @@ import RequirementsForm from './requirements-form/';
 import Focusable from './focusable/';
 import DefinitionList from './definition-list/';
 import Telephone from './telephone/';
+import Submit from './submit/';
 
 // Deprecated
 import UploadDroppable from './upload-droppable/';
@@ -35,5 +36,6 @@ export default angular.module('tw.styleguide.forms', [
   Focusable,
   DefinitionList,
   Telephone,
+  Submit,
   UploadDroppable
 ]).name;